mv vscode debugger input arg work-around to __init__
