+++ cli.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!\usr\bin\env python3
 
 from os import environ
 from pathlib import Path
@@ -19,91 +19,6 @@ SPIN_TXT_DATA = "Collecting Data from Aruba Central API Gateway..."
 tty = utils.tty
 
 app = typer.Typer()
-
-
-def get_arguments_from_import(import_file: str, key: str = None) -> list:
-    """Get arguments from default import_file (stored_tasks.yaml)
-
-    Args:
-        import_file (str): name of import file
-        key (str, optional): return single value for specific key if provided. Defaults to None.
-
-    Returns:
-        list: updated sys.argv list.
-    """
-    # args = utils.read_yaml(import_file)
-    args = config.get_config_data(Path(import_file))
-    if key and key in args:
-        args = args[key]
-
-    sys.argv += args
-
-    return sys.argv
-
-
-# -- break up arguments passed as single string from vscode promptString --
-def vscode_arg_handler():
-    try:
-        if len(sys.argv) > 1:
-            if " " in sys.argv[1] or not sys.argv[1]:
-                vsc_args = sys.argv.pop(1)
-                if vsc_args:
-                    if "\\'" in vsc_args:
-                        _loc = vsc_args.find("\\'")
-                        _before = vsc_args[:_loc - 1]
-                        _str_end = vsc_args.find("\\'", _loc + 1)
-                        sys.argv += _before.split()
-                        sys.argv += [f"{vsc_args[_loc + 2:_str_end]}"]
-                        sys.argv += vsc_args[_str_end + 2:].split()
-                    else:
-                        sys.argv += vsc_args.split()
-
-        if len(sys.argv) > 2:
-            _import_file, _import_key = None, None
-            if sys.argv[2].endswith((".yaml", ".yml", "json")):
-                _import_file = sys.argv.pop(2)
-                if not utils.valid_file(_import_file):
-                    if utils.valid_file(config.dir.joinpath(_import_file)):
-                        _import_file = config.dir.joinpath(_import_file)
-
-                if len(sys.argv) > 2:
-                    _import_key = sys.argv.pop(2)
-
-                sys.argv = get_arguments_from_import(_import_file, key=_import_key)
-
-    except Exception as e:
-        log.exception(f"Exception in vscode arg handler (arg split) {e.__class__}.{e}", show=True)
-        return
-
-    # update launch.json default if launched by vscode debugger
-    try:
-        launch_file = config.base_dir / ".vscode" / "launch.json"
-        launch_file_bak = config.base_dir / ".vscode" / "launch.json.bak"
-        if launch_file.is_file():
-            launch_data = launch_file.read_text()
-            launch_data = launch_data.splitlines()
-            for idx, line in enumerate(launch_data):
-                if "default" in line and "// VSC_PREV_ARGS" in line:
-                    _spaces = len(line) - len(line.lstrip(" "))
-                    new_line = f'{" ":{_spaces}}"default": "{" ".join(sys.argv[1:])}"  // VSC_PREV_ARGS'
-                    if line != new_line:
-                        log.debug(f"changing default arg for promptString:\n"
-                                  f"\t from: {line}\n"
-                                  f"\t to: {new_line}"
-                                  )
-                        launch_data[idx] = new_line
-
-                        # backup launch.json only if backup doesn't exist already
-                        if not launch_file_bak.is_file():
-                            import shutil
-                            shutil.copy(launch_file, launch_file_bak)
-
-                        # update launch.json
-                        launch_file.write_text("\n".join(launch_data))
-
-    except Exception as e:
-        log.exception(f"Exception in vscode arg handler (launch.json update) {e.__class__}.{e}", show=True)
-        return
 
 
 def eval_resp(resp):
@@ -493,11 +408,6 @@ def _refresh_tokens(account_name: str) -> CentralApi:
 
 
 # ---- // RUN \\ ----
-
-# TODO We may be able to do this in @app.callback which is run before the command
-# https://typer.tiangolo.com/tutorial/commands/context/
-if environ.get("TERM_PROGRAM") == "vscode":
-    vscode_arg_handler()
 
 # extract account from arguments or environment variables
 account = environ.get('ARUBACLI_ACCOUNT', "central_info")
